for aufs mmap: print the virtual file path
For details, read the document in later commit.
I don't like this approach, and UnionMount seems to try solving a
similar problem by customizing f_dentry and d_inode. Ideally it is
better to result the same fix, but it may not be so easy because
UnionMount doesn't have its virtual inode/dentry/file objects as it is
not a 'filesystem.'

Signed-off-by: J. R. Okajima <hooanon05g@gmail.com>